Fixed an issue with the NUSensorsActionators on 64bit machine. Added more noise in the fitness for webots.

+++ Infrastructure/NUData.cpp
@@ -157,10 +157,10 @@ string NUData::getStandardName(const string& hardwarename)
         if (currentletter.compare(string(" ")) != 0 && currentletter.compare(string("_")) != 0 && currentletter.compare(string("/")) != 0 && currentletter.compare(string("\\")) != 0 && currentletter.compare(string(".")) != 0)
             simplename += currentletter[0];            
     }
-    
+
     // Replace "Left"/"Right" with L/R and move to front of name
-    unsigned int Left = simplename.find("Left");
-    unsigned int Right = simplename.find("Right");
+    size_t Left = simplename.find("Left");
+    size_t Right = simplename.find("Right");
     if (Left != string::npos)
     {
         simplename.erase(Left, 4);
@@ -173,34 +173,34 @@ string NUData::getStandardName(const string& hardwarename)
     }
     
     // Replace plurals (ears, eyes)
-    unsigned int Ears = simplename.find("Ears");
-    unsigned int Eyes = simplename.find("Eyes");
+    size_t Ears = simplename.find("Ears");
+    size_t Eyes = simplename.find("Eyes");
     if (Ears != string::npos)
         simplename.replace(Ears, 4, "Ear");
     if (Eyes != string::npos)
         simplename.replace(Ears, 4, "Eye");
     
     // Replace ChestBoard with Chest
-    unsigned int ChestBoard = simplename.find("ChestBoard");
+    size_t ChestBoard = simplename.find("ChestBoard");
     if (ChestBoard != string::npos)
         simplename.replace(ChestBoard, 10, "Chest");
     
     // Replace LFace with LEye and RFace with REye
-    unsigned int LFace = simplename.find("LFace");
-    unsigned int RFace = simplename.find("RFace");
+    size_t LFace = simplename.find("LFace");
+    size_t RFace = simplename.find("RFace");
     if (LFace != string::npos)
     	simplename.replace(LFace, 5, "LEye");
     if (RFace != string::npos)
     	simplename.replace(RFace, 5, "REye");
     
     // Remove colours
-    unsigned int Red = simplename.find("Red");
+    size_t Red = simplename.find("Red");
     if (Red != string::npos)
         simplename.erase(Red, 3);
-    unsigned int Green = simplename.find("Green");
+    size_t Green = simplename.find("Green");
     if (Green != string::npos)
         simplename.erase(Green, 5);
-    unsigned int Blue = simplename.find("Blue");
+    size_t Blue = simplename.find("Blue");
     if (Blue != string::npos)
         simplename.erase(Blue, 4);
